[feat] : user관련 에러처리

+++ src/main/java/com/DreamOfDuck/talk/service/SessionService.java
@@ -1,5 +1,6 @@
 package com.DreamOfDuck.talk.service;
 
+import com.DreamOfDuck.account.entity.Member;
 import com.DreamOfDuck.global.exception.CustomException;
 import com.DreamOfDuck.global.exception.ErrorCode;
 import com.DreamOfDuck.talk.dto.request.SessionCreateRequest;
@@ -21,8 +22,7 @@ public class SessionService {
     private final SessionRepository sessionRepository;
 
     @Transactional
-    public SessionResponse save(SessionCreateRequest request){
-
+    public SessionResponse save(Member host, SessionCreateRequest request){
         Session session = Session.builder()
                 .duckType(Talker.fromValue(request.getDuckType()))
                 .isFormal(request.getIsFormal())
@@ -31,24 +31,38 @@ public class SessionService {
                         .collect(Collectors.toList()))
                 .cause(Cause.fromId(request.getCause()))
                 .build();
+        session.addHost(host);
         sessionRepository.save(session);
         return SessionResponse.from(session);
     }
     @Transactional
-    public SessionResponse update(SessionUpdateRequest request){
+    public SessionResponse update(Member host, SessionUpdateRequest request){
         Session session = sessionRepository.findById(request.getSessionId()).orElseThrow(EntityNotFoundException::new);
+       if(session.getHost()!=host){
+           throw new CustomException(ErrorCode.DIFFERENT_USER);
+       }
+
         session.setLast_emotion(LastEmotion.fromId(request.getLast_emotion()));
         if(session.getLast_emotion() == LastEmotion.OPTION4){
             if(request.getInput_field() ==null || request.getInput_field().isEmpty())  throw new CustomException(ErrorCode.EMPTY_INPUT_FIELD);
         }
+
         session.setInput_field(request.getInput_field());
         return SessionResponse.from(session);
     }
-    public SessionResponse findById(Long sessionId){
-        return SessionResponse.from(sessionRepository.findById(sessionId).orElseThrow(EntityNotFoundException::new));
+    public SessionResponse findById(Member host, Long sessionId){
+        Session session = sessionRepository.findById(sessionId).orElseThrow(EntityNotFoundException::new);
+        if(session.getHost()!=host){
+            throw new CustomException(ErrorCode.DIFFERENT_USER);
+        }
+        return SessionResponse.from(session);
     }
     @Transactional
-    public void delete(Long sessionId){
+    public void delete(Member host, Long sessionId){
+        Session session = sessionRepository.findById(sessionId).orElseThrow(EntityNotFoundException::new);
+        if(session.getHost()!=host){
+            throw new CustomException(ErrorCode.DIFFERENT_USER);
+        }
         sessionRepository.deleteById(sessionId);
     }
 }
